SecureMessenger/SecureMessenger/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.5",
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "oauthlib>=3.3.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
    "wtforms>=3.2.1",
    "flask-wtf>=1.2.2",
]
